* origin/master: (81 commits)
  Remove Clang warnings
  Try to fix Jenkins Windows builds
  Turn Configuration keys from std::string to constexpr const char* for better startup performance
  Support stb_image based texture reader (closes #541)
  Apply coding styles
  Introduce property to toggle dashboard globally (closes #545)
  Include the build id into the Jenkins path
  Undo change in globebrowsing.asset file
  Undo change in World_Imagery file
  Update Ghoul repository
  Feature/new formatting (#547)
  One Property Tree (#500)
  Cleanup FieldlineSequence classes
  Fix crash if screenshot folder already existed
  Include additional performance measurement hooks into the Framebufferenderer
  Feature/jenkins warnings (#538)
  Update Ghoul repository
  Fix module specification
  Add new dedicated directory for screenshots
  Reuse the same font renderer program object between DU rendering classes
  ...

# Conflicts:
#	Jenkinsfile
#	modules/multiresvolume/rendering/histogrammanager.cpp
#	modules/multiresvolume/rendering/localerrorhistogrammanager.cpp
#	modules/multiresvolume/rendering/renderablemultiresvolume.cpp
#	modules/multiresvolume/rendering/tsp.cpp

+++ modules/touch/include/tuioear.h
@@ -107,7 +107,7 @@ class TuioEar : public TUIO::TuioListener {
         /**
         * A list that tracks all of the cursor ID's that got removed since last frame
         */
-        std::vector<int> _removeList;
+        std::vector<long> _removeList;
 };
 
 #endif // __OPENSPACE_MODULE_TOUCH___TUIO_EAR___H__
